requirements.txt
django-filter==1.1.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
Django==2.0.6
django-heroku==0.3.1
gunicorn==19.9.0
django-cors-headers==2.2.0
pytz==2018.9
psycopg2-binary==2.7.7
django-allauth==0.36.0